skalemon/requirements.txt
requests
prometheus-client
pyyaml

